Refactor binaries a little
